Port the test for Undefinable.andThen() to ava
